+++ Editor/UserSettings.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using UnityEngine.UIElements;
 
 namespace UnityEditor.U2D.Aseprite
 {
@@ -6,6 +7,7 @@ namespace UnityEditor.U2D.Aseprite
     {
         const bool k_DefaultBackgroundImport = false;
         const string k_BackgroundImportKey = UserSettings.settingsUniqueKey + "ImportSettings.backgroundImport";
+        const string k_SettingsTitle = "Aseprite Importer";
         static readonly GUIContent k_BackgroundImportLabel = EditorGUIUtility.TrTextContent("Background import", "Enable asset import when the Unity Editor is in the background.");
 
         public static bool backgroundImport
@@ -14,19 +16,35 @@ namespace UnityEditor.U2D.Aseprite
             private set => EditorPrefs.SetBool(k_BackgroundImportKey, value);
         }
 
-        public void OnGUI()
+        public static void SetupUI(VisualElement rootElement)
         {
-            EditorGUI.BeginChangeCheck();
-            var c = EditorGUILayout.Toggle(k_BackgroundImportLabel, backgroundImport);
-            if (EditorGUI.EndChangeCheck())
-                backgroundImport = c;
+            var container = new VisualElement();
+            container.style.paddingLeft = 5;
+            rootElement.Add(container);
+            
+            var header = new Label(k_SettingsTitle);
+            header.style.paddingLeft = 5;
+            header.style.paddingBottom = 10;
+            header.style.fontSize = 20;
+            header.style.unityFontStyleAndWeight = new StyleEnum<FontStyle>(FontStyle.Bold);
+            container.Add(header);
+            
+            var toggle = new Toggle(k_BackgroundImportLabel.text)
+            {
+                tooltip = k_BackgroundImportLabel.tooltip,
+                value = backgroundImport
+            };
+            toggle.RegisterValueChangedCallback(x =>
+            {
+                backgroundImport = x.newValue;
+            });
+            container.Add(toggle);
         }
     }
 
     internal class UserSettings : SettingsProvider
     {
         public const string settingsUniqueKey = "UnityEditor.U2D.Aseprite/";
-        static readonly ImportSettings s_ImportSettings = new ImportSettings();
 
         UserSettings() : base("Preferences/2D/Aseprite Importer", SettingsScope.User)
         {
@@ -34,17 +52,11 @@ namespace UnityEditor.U2D.Aseprite
         }
 
         [SettingsProvider]
-        static SettingsProvider CreateSettingsProvider()
-        {
-            return new UserSettings()
-            {
-                guiHandler = SettingsGUI
-            };
-        }
+        static SettingsProvider CreateSettingsProvider() { return new UserSettings(); }
 
-        static void SettingsGUI(string searchContext)
+        public override void OnActivate(string searchContext, VisualElement rootElement)
         {
-            s_ImportSettings.OnGUI();
+            ImportSettings.SetupUI(rootElement);
         }
     }
 }